episodes/episode-001/requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.2
